requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
boto3==1.34.0
click==8.2.1
fastapi==0.116.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
h11==0.16.0
httpx==0.25.2
idna==3.10
motor==3.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
